+++ tests/unit_tests/features/test_decorators.py
@@ -1,4 +1,4 @@
-def test_bivariate_feature_undirected():
+def test_bivariate_feature():
     """Test bivariate_feature creates BivariateFeature (not directed)."""
     from eegdash.features.decorators import BivariateFeature, bivariate_feature
 
@@ -6,33 +6,11 @@ def test_bivariate_feature_undirected():
         return x + y
 
     # Apply decorator with directed=False (line 132)
-    decorated = bivariate_feature(test_func, directed=False)
+    decorated = bivariate_feature(test_func)
 
     # Check that the function has been decorated
     assert hasattr(decorated, "feature_kind")
     assert isinstance(decorated.feature_kind, BivariateFeature)
-
-
-def test_bivariate_feature_directed():
-    """Test bivariate_feature with directed=True (line 132)."""
-    from eegdash.features.decorators import bivariate_feature
-    from eegdash.features.extractors import DirectedBivariateFeature
-
-    @bivariate_feature
-    def dummy_undirected(x):
-        return x
-
-    @bivariate_feature
-    def dummy_directed(x):
-        return x
-
-    # Apply with directed=True
-    dummy_directed_applied = bivariate_feature(lambda x: x, directed=True)
-
-    from eegdash.features.extractors import _get_underlying_func
-
-    kind = _get_underlying_func(dummy_directed_applied).feature_kind
-    assert isinstance(kind, DirectedBivariateFeature)
 
 
 def test_feature_predecessor_empty():
@@ -43,15 +21,15 @@ def test_feature_predecessor_empty():
     def my_func(x):
         return x
 
-    from eegdash.features.extractors import _get_underlying_func
+    from eegdash.features.base_utils import get_underlying_func
 
-    assert _get_underlying_func(my_func).parent_extractor_type == [None]
+    assert get_underlying_func(my_func).parent_extractor_type == [None]
 
 
 def test_feature_kind_decorator():
     """Test FeatureKind decorator."""
     from eegdash.features.decorators import FeatureKind
-    from eegdash.features.extractors import UnivariateFeature
+    from eegdash.features.kinds import UnivariateFeature
 
     kind_instance = UnivariateFeature()
 
@@ -59,6 +37,6 @@ def test_feature_kind_decorator():
     def my_feature(x):
         return x
 
-    from eegdash.features.extractors import _get_underlying_func
+    from eegdash.features.base_utils import get_underlying_func
 
-    assert _get_underlying_func(my_feature).feature_kind is kind_instance
+    assert get_underlying_func(my_feature).feature_kind is kind_instance
