+++ tests/unit_tests/features/test_utils.py
@@ -13,7 +13,7 @@ from eegdash.features.utils import (
 
 
 def test_utils_gaps(features_dataset):
-    from eegdash.features.extractors import _get_underlying_func
+    from eegdash.features.base_utils import get_underlying_func
 
     # 52-85: extract_features
     mock_ds = MagicMock()
@@ -59,7 +59,7 @@ def test_utils_gaps(features_dataset):
     fit_feature_extractors(mock_ds, {"f": lambda x: x})
 
     # TrainableFeature gaps (62, 76)
-    from eegdash.features.extractors import TrainableFeature
+    from eegdash.features.trainable import TrainableFeature
 
     class MyRawTrainable(TrainableFeature):
         def clear(self):
@@ -83,9 +83,8 @@ def test_utils_gaps(features_dataset):
     # Let's check extractors.py again.
 
     # 295-302: MultivariateFeature.__call__ errors/edge cases
-    from eegdash.features.extractors import (
-        BivariateFeature,
-        DirectedBivariateFeature,
+    from eegdash.features.base_utils import BivariateIterator
+    from eegdash.features.kinds import (
         MultivariateFeature,
     )
 
@@ -95,11 +94,11 @@ def test_utils_gaps(features_dataset):
     mv._array_to_dict(np.array([1]), [])
 
     # 361, 377: pair indices
-    BivariateFeature.get_pair_iterators(3)
-    DirectedBivariateFeature.get_pair_iterators(3)
+    BivariateIterator(3).get_pair_iterators()
+    BivariateIterator(3, directed=True).get_pair_iterators()
 
     # 178, 183-192: fit_feature_extractors
-    from eegdash.features.extractors import UnivariateFeature
+    from eegdash.features.kinds import UnivariateFeature
 
     class MyTrainable(FeatureExtractor):
         def __init__(self, *args, **kwargs):
@@ -167,7 +166,7 @@ def test_utils_gaps(features_dataset):
 
     mock_dispatcher = MagicMock(spec=Dispatcher)
     mock_dispatcher.py_func = lambda x: x
-    _get_underlying_func(mock_dispatcher)
+    get_underlying_func(mock_dispatcher)
 
     # 205: call trainable fe
     trainable_fe(
@@ -197,14 +196,14 @@ def test_utils_gaps(features_dataset):
     fe_outer.fit()
 
     # 340, 365: feature_channel_names
-    from eegdash.features.extractors import BivariateFeature, DirectedBivariateFeature
+    from eegdash.features.kinds import BivariateFeature
 
     UnivariateFeature().feature_channel_names(_metadata={"info": {"ch_names": ["ch1"]}})
     BivariateFeature().feature_channel_names(
-        _metadata={"info": {"ch_names": ["ch1", "ch2"]}}
-    )
-    DirectedBivariateFeature().feature_channel_names(
-        _metadata={"info": {"ch_names": ["ch1", "ch2"]}}
+        _metadata={
+            "info": {"ch_names": ["ch1", "ch2"]},
+            "ch_pair_iterator": BivariateIterator(2),
+        }
     )
 
     pass
@@ -214,14 +213,14 @@ def test_extract_features_list_input():
     """Test extract_features with list of feature functions."""
     import numpy as np
 
-    from eegdash.features import extractors
+    from eegdash.features import kinds
 
     # Create a simple feature function
     def mean_feature(x):
         return np.mean(x, axis=-1, keepdims=True)
 
     mean_feature.parent_extractor_type = [None]
-    mean_feature.feature_kind = extractors.UnivariateFeature()
+    mean_feature.feature_kind = kinds.UnivariateFeature()
 
     # Test that list conversion works
     # This tests line 132 and 134
